Button for launching ocr tests in ResultActivity
