+++ app/src/main/res/layout/activity_result.xml
@@ -41,15 +41,24 @@
         <!--</ScrollView>-->
     </LinearLayout>
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/newPictureFab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:src="@drawable/ic_photo_camera_24px"
-        android:layout_margin="16dp"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:backgroundTint="@color/colorPrimary"/>
 
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <Button
+            android:id="@+id/test_button"
+            android:text="@string/test"
+            android:layout_height="wrap_content"
+            android:layout_width="wrap_content"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentEnd="true"/>
+        <Button
+            android:id="@+id/newPictureFab"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/pic"
+            android:layout_above="@+id/test_button"
+            android:layout_alignParentEnd="true" />
+    </RelativeLayout>
 </android.support.constraint.ConstraintLayout>
